Update warnings unity 5. Everyplay deprecations cleanup.  Update points to double.

+++ Gameverses/Community/UI/GameCommunityUIPanelAchievements.cs
@@ -72,7 +72,7 @@ public class GameCommunityUIPanelAchievements : UIAppPanelBaseList {
             
             int i = 0;
             
-            int totalPoints = 0;
+            double totalPoints = 0;
             
             foreach (GameAchievement achievement in achievements) {
                 
@@ -111,7 +111,7 @@ public class GameCommunityUIPanelAchievements : UIAppPanelBaseList {
                 string points = "";
                 
                 if (completed) {
-                    int currentPoints = achievement.data.points;
+                    double currentPoints = achievement.data.points;
                     totalPoints += currentPoints;
                     points = currentPoints.ToString();
